move Timerange to a separate class
