+++ planet/src/main/webapp/itinerary.html
@@ -24,7 +24,7 @@
         <!-- Include the firebase and local JS files -->  
         <script src="https://www.gstatic.com/firebasejs/4.1.3/firebase.js"></script> 
         <script src="js/authentication.js"></script>
-        <script src="js/itinerary.js"></script>
+        <script type="module" src="js/itinerary.js"></script>
 
         <!--Let browser know website is optimized for mobile-->    
         <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
